修改商品主键为bigint，订单主键为bigint

+++ biz/src/main/java/com/bytecamp/biz/service/ProductService.java
@@ -13,5 +13,5 @@ public interface ProductService {
      * @param id 商品id
      * @return
      */
-    Product getProductById(Integer id);
+    Product getProductById(Long id);
 }
